test to see if the file exists

+++ get_enhanced-media_war.sh
@@ -1,10 +1,8 @@
 #!/bin/bash
-artifact=mediaserver-nf.ear
+FILE=mediaserver-nf.ear
+SRC=/media/buffalo/dina-data/naturalist/integrated/${FILE}
 DST=srv/deployments
 RELEASE=srv/releases
 
-echo "fetching the artifact ${artifact} from Buffalo"
-SRC=/media/buffalo/dina-data/naturalist/integrated/${artifact}
 echo "copying from ${SRC} to ${RELEASE} and ${DST}"
-cp $SRC $RELEASE
-cp $SRC $DST
+test -f $DST/$FILE || cp $SRC $RELEASE && cp $SRC $DST
